add error handling

+++ src/main/java/com/heroku/nwl/service/CallbackQueryHandler.java
@@ -2,13 +2,16 @@ package com.heroku.nwl.service;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.heroku.nwl.config.CustomBotException;
 import com.heroku.nwl.constants.Constants;
 import com.heroku.nwl.dto.ButtonDto;
 import com.heroku.nwl.dto.CalendarDayDto;
 import com.heroku.nwl.model.DayOff;
 import com.heroku.nwl.model.DayOffRepository;
-import com.heroku.nwl.model.ReservationRepository;
 import com.heroku.nwl.model.Reservation;
+import com.heroku.nwl.model.ReservationRepository;
+import com.heroku.nwl.model.ReservationStatus;
+import com.heroku.nwl.model.Role;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
@@ -23,7 +26,7 @@ import java.time.format.DateTimeFormatter;
 import java.util.List;
 import java.util.Objects;
 
-import static com.heroku.nwl.constants.Commands.ALL_RESERVATION_ON_DATE;
+import static com.heroku.nwl.constants.Commands.ALL_RESERVATION;
 import static com.heroku.nwl.constants.Commands.AVAILABLE_DATE_TO_RESERVE;
 import static com.heroku.nwl.constants.Commands.CANCEL_RESERVE;
 import static com.heroku.nwl.constants.Commands.CHANGE_MONTH;
@@ -32,15 +35,20 @@ import static com.heroku.nwl.constants.Commands.GO_BACK;
 import static com.heroku.nwl.constants.Commands.ORDER_TIME;
 import static com.heroku.nwl.constants.Commands.SELECT_SERVICE;
 import static com.heroku.nwl.constants.Commands.WORKDAY;
+import static com.heroku.nwl.constants.Constants.ACTIVE_BOOKING;
+import static com.heroku.nwl.constants.Constants.ADD_DAY_OFF;
 import static com.heroku.nwl.constants.Constants.CHOOSE_DATE;
 import static com.heroku.nwl.constants.Constants.CHOOSE_DAY_OFF;
 import static com.heroku.nwl.constants.Constants.CHOOSE_SERVICE;
 import static com.heroku.nwl.constants.Constants.DATE_PATTERN;
+import static com.heroku.nwl.constants.Constants.DELETE_DAY_OFF;
 import static com.heroku.nwl.constants.Constants.DELETE_RESERVATION_MESSAGE;
 import static com.heroku.nwl.constants.Constants.ERROR;
 import static com.heroku.nwl.constants.Constants.ERROR_DELETE_RESERVATION_MESSAGE;
-import static com.heroku.nwl.service.Calendar.ADD_DAY_OFF;
-import static com.heroku.nwl.service.Calendar.DELETE_DAY_OFF;
+import static com.heroku.nwl.constants.Constants.TIME_ALREADY_TAKEN;
+import static com.heroku.nwl.constants.Constants.USER_NEW_RESERVATION_MESSAGE;
+import static com.heroku.nwl.constants.ErrorMessage.ERROR_PERMISSION;
+
 
 
 
@@ -52,8 +60,9 @@ public class CallbackQueryHandler {
     private final DayOffRepository dayOffRepository;
     private final KeyboardService keyboardService;
     private final ReservationRepository reservationRepository;
+    private final UserService userService;
 
-    public EditMessageText getEditMessage(Update update) {
+    public EditMessageText getEditMessage(Update update) throws CustomBotException {
         EditMessageText message;
         String callbackData = update.getCallbackQuery().getData();
         String text = Constants.INCORRECT_DATE;
@@ -67,11 +76,16 @@ public class CallbackQueryHandler {
         }
         message = switch (buttonDto.getCommand()) {
             case CHANGE_MONTH -> getChangeMonthMessage(buttonDto, chatId, messageId);
-            case ALL_RESERVATION_ON_DATE -> prepareEditMessageText(
-                    CHOOSE_DATE,
-                    chatId,
-                    messageId,
-                    keyboardService.getReservationOnDate(buttonDto.getCurrentDate()));
+            case ALL_RESERVATION -> {
+                if (!userService.getUserRole(chatId).equals(Role.ADMIN)) {
+                    throw new CustomBotException(ERROR_PERMISSION);
+                }
+                yield prepareEditMessageText(
+                        CHOOSE_DATE,
+                        chatId,
+                        messageId,
+                        keyboardService.getReservationOnDate(buttonDto.getCurrentDate()));
+            }
             case ADD_DAY_OFF -> getAddDayOffMessage(buttonDto, text, chatId, messageId);
             case DELETE_DAY_OFF -> getDeleteDayOffMessage(buttonDto, text, chatId, messageId);
             case WORKDAY -> getWorkDayMessage(buttonDto, chatId, messageId);
@@ -85,20 +99,38 @@ public class CallbackQueryHandler {
         return message;
     }
 
-    private EditMessageText getChangeMonthMessage(ButtonDto buttonDto, Long chatId, Long messageId) {
+    private EditMessageText getChangeMonthMessage(ButtonDto buttonDto, Long chatId, Long messageId) throws CustomBotException {
+        if (!userService.getUserRole(chatId).equals(Role.ADMIN)) {
+            throw new CustomBotException(ERROR_PERMISSION);
+        }
         List<List<CalendarDayDto>> calendar = calendarService.getCalendar(buttonDto.getCurrentDate(), buttonDto.getReturnTo());
         InlineKeyboardMarkup keyboardMarkup = keyboardService.getCalendar(calendar, buttonDto.getCurrentDate());
         return prepareEditMessageText(CHOOSE_DATE, chatId, messageId, keyboardMarkup);
     }
 
-    private EditMessageText getCreateCalendarMessage(ButtonDto buttonDto, String text, long chatId, long messageId) {
+    private EditMessageText getCreateCalendarMessage(ButtonDto buttonDto, String text, long chatId, long messageId) throws CustomBotException {
+        if (!userService.getUserRole(chatId).equals(Role.ADMIN)) {
+            throw new CustomBotException(ERROR_PERMISSION);
+        }
         List<List<CalendarDayDto>> calendar = calendarService.getCalendar(buttonDto.getCurrentDate(), ADD_DAY_OFF);
         InlineKeyboardMarkup keyboardMarkup = keyboardService.getCalendar(calendar, buttonDto.getCurrentDate());
         return prepareEditMessageText(text, chatId, messageId, keyboardMarkup);
     }
 
-    private EditMessageText getAddDayOffMessage(ButtonDto buttonDto, String text, long chatId, long messageId) {
+    private EditMessageText getAddDayOffMessage(ButtonDto buttonDto, String text, long chatId, long messageId) throws CustomBotException {
+        if (!userService.getUserRole(chatId).equals(Role.ADMIN)) {
+            throw new CustomBotException(ERROR_PERMISSION);
+        }
         DayOff dayOff = new DayOff(buttonDto.getCurrentDate());
+        List<Reservation> reservationList = reservationService.getReservationsByDate(dayOff.getDayOffDate());
+        if(!reservationList.isEmpty()){
+            for (Reservation reservation:reservationList
+                 ) {
+                if(reservation.getReservationStatus().equals(ReservationStatus.ACTIVE)){
+                    return getCreateCalendarMessage(buttonDto, ACTIVE_BOOKING, chatId, messageId);
+                }
+            }
+        }
         if (LocalDate.now().isBefore(dayOff.getDayOffDate())) {
             dayOffRepository.save(dayOff);
             text = CHOOSE_DAY_OFF;
@@ -106,14 +138,14 @@ public class CallbackQueryHandler {
         return getCreateCalendarMessage(buttonDto, text, chatId, messageId);
     }
 
-    private EditMessageText getReservationOnTimeMessage(ButtonDto buttonDto, long chatId, long messageId) {
+    private EditMessageText getReservationOnTimeMessage(ButtonDto buttonDto, long chatId, long messageId) throws CustomBotException {
         EditMessageText message;
         LocalDate orderDate = buttonDto.getCurrentDate();
         LocalTime orderTime = buttonDto.getReservedTime();
         if (reservationService.createReservation(orderTime, orderDate, chatId,buttonDto.getServiceId())) {
-            message = prepareEditMessageText("You are reserve " + orderDate + " " + orderTime, chatId, messageId, null);
+            message = prepareEditMessageText(String.format(USER_NEW_RESERVATION_MESSAGE,orderDate, orderTime), chatId, messageId, null);
         } else {
-            message = prepareEditMessageText("Sorry this time is already taken or you don't register", chatId, messageId, null);
+            message = prepareEditMessageText(TIME_ALREADY_TAKEN, chatId, messageId, null);
         }
         return message;
     }
@@ -133,7 +165,10 @@ public class CallbackQueryHandler {
                 keyboardService.getUserReservation(chatId));
     }
 
-    private EditMessageText getDeleteDayOffMessage(ButtonDto buttonDto, String text, long chatId, long messageId) {
+    private EditMessageText getDeleteDayOffMessage(ButtonDto buttonDto, String text, long chatId, long messageId) throws CustomBotException {
+        if (!userService.getUserRole(chatId).equals(Role.ADMIN)) {
+            throw new CustomBotException(ERROR_PERMISSION);
+        }
         DayOff dayOff = new DayOff(buttonDto.getCurrentDate());
         if (LocalDate.now().isBefore(dayOff.getDayOffDate())) {
             dayOffRepository.delete(dayOff);
@@ -142,7 +177,7 @@ public class CallbackQueryHandler {
         return getCreateCalendarMessage(buttonDto, text, chatId, messageId);
     }
 
-    private EditMessageText getWorkDayMessage(ButtonDto buttonDto, long chatId, long messageId) {
+    private EditMessageText getWorkDayMessage(ButtonDto buttonDto, long chatId, long messageId) throws CustomBotException {
         DateTimeFormatter dateTimeFormatter = DateTimeFormatter.ofPattern(DATE_PATTERN);
         LocalDate currentDate = buttonDto.getCurrentDate();
         String formattedDate = currentDate.getDayOfWeek() + "\n" + currentDate.format(dateTimeFormatter);
@@ -152,7 +187,8 @@ public class CallbackQueryHandler {
                 messageId,
                 keyboardService.getAvailableTimeKeyboard(buttonDto));
     }
-    private EditMessageText getServiceCatalog(ButtonDto buttonDto,  long chatId, long messageId){
+
+    private EditMessageText getServiceCatalog(ButtonDto buttonDto, long chatId, long messageId) throws CustomBotException {
         return prepareEditMessageText(
                 CHOOSE_SERVICE,
                 chatId,
@@ -165,24 +201,33 @@ public class CallbackQueryHandler {
         if (Objects.equals(buttonDto.getReturnTo(), AVAILABLE_DATE_TO_RESERVE)) {
             message = prepareEditMessageText(CHOOSE_DATE, chatId, messageId, keyboardService.getScheduleDays());
         }
-        if (buttonDto.getReturnTo().equals(ALL_RESERVATION_ON_DATE)) {
-            List<List<CalendarDayDto>> calendar = calendarService.getCalendar(buttonDto.getCurrentDate(), ALL_RESERVATION_ON_DATE);
+        if (buttonDto.getReturnTo().equals(ALL_RESERVATION)) {
+            List<List<CalendarDayDto>> calendar = calendarService.getCalendar(buttonDto.getCurrentDate(), ALL_RESERVATION);
             InlineKeyboardMarkup keyboardMarkup = keyboardService.getCalendar(calendar, buttonDto.getCurrentDate());
             message = prepareEditMessageText(CHOOSE_DATE, chatId, messageId, keyboardMarkup);
         }
-
         return message;
     }
 
     @Transactional(propagation = Propagation.REQUIRES_NEW)
-    public EditMessageText canselReservation(ButtonDto buttonDto, Long chatId, Long messageId) {
+    public EditMessageText canselReservation(ButtonDto buttonDto, Long chatId, Long messageId) throws CustomBotException {
+        if (!userService.getUserRole(chatId).equals(Role.ADMIN)) {
+            throw new CustomBotException(ERROR_PERMISSION);
+        }
         Reservation order = reservationRepository.findByOrderId(buttonDto.getReservationId());
         reservationService.canselReservation(buttonDto.getReservationId());
-        String text = "You are cansel reserve " + order.getOrderTime() + " " + order.getOrderDate();
-        return prepareEditMessageText(text, chatId, messageId, keyboardService.getReservationOnDate(buttonDto.getCurrentDate()));
+        String text = String.format(Constants.MESSAGE_USER_CANSEL_RESERVE, order.getOrderTime(), order.getOrderDate());
+        return prepareEditMessageText(
+                text,
+                chatId,
+                messageId,
+                keyboardService.getReservationOnDate(buttonDto.getCurrentDate()));
     }
 
-    private EditMessageText prepareEditMessageText(String text, long chatId, long messageId, InlineKeyboardMarkup inlineKeyboardMarkup) {
+    private EditMessageText prepareEditMessageText(String text,
+                                                   long chatId,
+                                                   long messageId,
+                                                   InlineKeyboardMarkup inlineKeyboardMarkup) {
         EditMessageText message = new EditMessageText();
         message.setChatId(String.valueOf(chatId));
         message.setText(text);
